requirements.txt
uvicorn[standard]==0.20.0
fastapi==0.92.0
loguru==0.6.0
aiohttp==3.8.4
redis==4.2.0rc1
motor==3.1.1
pytz==2022.7.1
PyJWT==2.6.0
pydantic==1.10.6
starlette==0.25.0
cryptography==39.0.2
pymongo~=4.3.3
httpx~=0.23.3
APScheduler~=3.10.1
aiofiles==23.1.0